backend/requirements.txt
alembic==1.13.1
argcomplete==3.2.2
bcrypt==4.1.2
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
distlib==0.3.8
exceptiongroup==1.2.1
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mutagen==1.47.0
mypy==1.9.0
mypy-extensions==1.0.0
nox==2024.4.15
packaging==23.2
platformdirs==4.2.0
pluggy==1.5.0
psycopg2==2.9.9
pycryptodomex==3.20.0
pytest==8.1.1
pytest-flask==1.3.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.27
tomli==2.0.1
types-Flask-Cors==4.0.0.20240405
typing_extensions==4.9.0
urllib3==2.2.0
virtualenv==20.25.0
websockets==12.0
Werkzeug==3.0.1
yt-dlp==2024.4.9
